@@ -1,11 +1,11 @@
 plugins {
     id 'com.android.application'
     id 'org.jetbrains.kotlin.android'
-    id 'kotlin-android-extensions'
     id 'kotlin-kapt'
     id 'com.google.gms.google-services'
     id 'com.google.firebase.crashlytics'
     id 'com.google.android.libraries.mapsplatform.secrets-gradle-plugin'
+    id 'kotlin-parcelize'
 }
 
 android {
@@ -67,12 +67,12 @@ android {
     }
 
     compileOptions {
-        sourceCompatibility JavaVersion.VERSION_1_8
-        targetCompatibility JavaVersion.VERSION_1_8
+        sourceCompatibility JavaVersion.VERSION_17
+        targetCompatibility JavaVersion.VERSION_17
     }
 
     kotlinOptions {
-        jvmTarget = '1.8'
+        jvmTarget = '17'
     }
 
     flavorDimensions 'driver'
@@ -146,6 +146,7 @@ android {
             }
         }
     }
+    namespace 'com.logestechs.driver'
 
 }
 
@@ -170,7 +171,7 @@ dependencies {
     implementation "com.github.YarikSOffice:lingver:1.3.0"
 
     //Shared Pref wrapper
-    implementation 'com.cioccarellia:ksprefs:2.2.1'
+    implementation 'com.github.cioccarellia:ksprefs:2.4.1'
     implementation 'androidx.swiperefreshlayout:swiperefreshlayout:1.1.0'
 
     //Image Compression
@@ -201,10 +202,10 @@ dependencies {
     androidTestImplementation 'androidx.test.ext:junit:1.1.3'
     androidTestImplementation 'androidx.test.espresso:espresso-core:3.4.0'
 
-    implementation 'com.synnapps:carouselview:0.1.6'
+    //implementation 'com.synapse:carouselview:0.1.5'
 
     //SliderView
-    implementation 'com.github.smarteist:autoimageslider:1.3.9'
+    implementation 'com.github.smarteist:Android-Image-Slider:1.4.0'
     implementation "com.github.bumptech.glide:glide:4.11.0"
     //ZoomImages
     implementation 'com.github.chrisbanes:PhotoView:2.3.0'
